requirements.txt
requests
pyodbc
azure-storage-blob
dotenv
PyPDF2
pdfplumber
openai
pandas
azure.search.documents
fastapi
uvicorn